Refactored coupled_global
