+++ zppy_interfaces/global_time_series/coupled_global_utils.py
@@ -1,0 +1,37 @@
+from enum import Enum
+
+
+class Metric(Enum):
+    AVERAGE = 1
+    TOTAL = 2
+
+
+class Variable(object):
+    def __init__(
+        self,
+        variable_name,
+        metric=Metric.AVERAGE,
+        scale_factor=1.0,
+        original_units="",
+        final_units="",
+        group="All Variables",
+        long_name="",
+    ):
+        # The name of the EAM/ELM/etc. variable on the monthly h0 history file
+        self.variable_name: str = variable_name
+
+        # These fields are used for computation
+        # Global average over land area or global total
+        self.metric: Metric = metric
+        # The factor that should convert from original_units to final_units, after standard processing with nco
+        self.scale_factor: float = scale_factor
+        # Test string for the units as given on the history file (included here for possible testing)
+        self.original_units: str = original_units
+        # The units that should be reported in time series plots, based on metric and scale_factor
+        self.final_units: str = final_units
+
+        # These fields are used for plotting
+        # A name used to cluster variables together, to be separated in groups within the output web pages
+        self.group: str = group
+        # Descriptive text to add to the plot page to help users identify the variable
+        self.long_name: str = long_name
