+++ zppy_interfaces/global_time_series/index_template.html
