+++ zppy_interfaces/multi_utils/logger.py
@@ -17,6 +17,8 @@ console_handler.setFormatter(logFormatter)
 logging.getLogger().addHandler(console_handler)
 
 
+# TODO: Why are logging messages printing 3x?
+# Following the example of https://github.com/xCDAT/xcdat/blob/27396e510185fddfe04921427ce2d67618a93709/xcdat/_logger.py#L20
 def _setup_custom_logger(name, propagate=True) -> logging.Logger:
     """Sets up a custom logger.
 
@@ -27,7 +29,7 @@ def _setup_custom_logger(name, propagate=True) -> logging.Logger:
     name : str
         Name of the file where this function is called.
     propagate : bool, optional
-        Whether to propagate logger messages or not, by default False
+        Whether to propagate logger messages or not, by default True
 
     Returns
     -------
@@ -63,6 +65,6 @@ def _setup_custom_logger(name, propagate=True) -> logging.Logger:
     """
     logger = logging.getLogger(name)
     logger.propagate = propagate
-    logger.setLevel(logging.DEBUG)
+    logger.setLevel("DEBUG")
 
     return logger
